requirements.txt
python-dotenv==0.21.0
discord.py==2.1.0
twilio==7.16.0
